Repaired StackOverflow: null caused by ObjectDTO where another object was called from relation

+++ src/main/java/com/example/task_manager/tasks/controllers/TaskController.java
@@ -20,7 +20,7 @@ public class TaskController {
 
 
     @GetMapping("/list/tasks/")
-    public ResponseEntity<Set<ProtectedTaskDTO>> getAllTasks(int project) {
+    public ResponseEntity<Set<ProtectedTaskDTO>> getAllTasks(Long project) {
         return taskService.getAllTasks(project);
     }
     @GetMapping("/find/task/")
